iOS-11: cast value to int before sending to controller (controller assumes the value is an int)

+++ ORControllerClient/ORControllerClient/ORController.m
@@ -408,7 +408,7 @@
 
 - (void)sendValue:(float)value forSlider:(ORSlider *)sender
 {
-    [self controlForWidget:sender action:[NSString stringWithFormat:@"%f", value]];
+    [self controlForWidget:sender action:[NSString stringWithFormat:@"%d", (int)value]];
 }
 
 - (void)sendColor:(UIColor *)color forPicker:(ORColorPicker *)sender
